requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
langchain-openai
langchain-google-genai
langchain-community
langchain-pinecone
duckduckgo-search
pinecone
supabase
pydantic
langsmith
amadeus
fast-flights
httpx
streamlit
langchainhub
isodate
kagglehub
kaggle
pandas